Edit expansion and parsing
