+++ sources/parser/handle_cmd_args_utils.c
@@ -6,7 +6,7 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 18:12:21 by henbuska          #+#    #+#             */
-/*   Updated: 2024/12/20 17:43:10 by henbuska         ###   ########.fr       */
+/*   Updated: 2024/12/21 16:25:22 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,116 +19,9 @@ int	skip_whitespace(char *str, int i)
 	return (i);
 }
 
-int	arg_in_quotes(t_shell *mini, char *str, int i, t_expand *arg)
+int	arg_no_quotes(t_cmd *cmd, t_expand *arg, int i)
 {
-	int j = 0;
-	(void)mini;
-	if (the_arg(arg, i))
-		return (-1);
-	what_quote(str, arg);
-	while (str[arg->i])
-	{
-		if (str[arg->i] == '$' && str[arg->i + 1] && (str[arg->i + 1] == '\'' || str[arg->i + 1] == '"') && (!arg->sgl && !arg->dbl))
-			arg->i++;
-		else if (!arg->sgl && !arg->dbl && (str[arg->i] == '\'' || str[arg->i] == '"'))
-			what_quote(str, arg);
-		else if ((arg->sgl && str[arg->i] == '\'') || (arg->dbl && str[arg->i] == '"'))
-			what_quote(str, arg);
-		else if (add_char(str, arg))
-			return (free(arg->value), -1);
-		if (str[arg->i] == ' ' && !arg->sgl && !arg->dbl)
-			break ;
-		j++;
-	}
-	arg->len = ft_strlen(arg->value);
-	return (arg->i);
-}
-
-int	segment_in_quotes(t_shell *mini, char *str, int i, t_expand *arg)
-{
-	int j = 0;
-	if (the_arg(arg, i))
-		return (-1);
-	what_quote(str, arg);
-	while (str[arg->i])
-	{
-		if (!arg->sgl && !arg->dbl)
-			break ;
-		if (str[arg->i] == '$' && str[arg->i + 1] && (str[arg->i + 1] == '\'' || str[arg->i + 1] == '"') && (!arg->sgl && !arg->dbl))
-			arg->i++;
-		if ((str[arg->i] == '$' && !arg->sgl) && (str[arg->i + 1] && !ft_isspace(str[arg->i + 1])
-			&& (ft_isalnum(str[arg->i + 1]) || str[arg->i + 1] == '_' || str[arg->i + 1] == '?')))
-		{
-			if (we_have_dollar(mini, arg, str) == -1)
-				return (free(arg->value), -1);
-		}
-		else if (!arg->sgl && !arg->dbl && (str[arg->i] == '\'' || str[arg->i] == '"'))
-			what_quote(str, arg);
-		else if ((arg->sgl && str[arg->i] == '\'') || (arg->dbl && str[arg->i] == '"'))
-			what_quote(str, arg);
-		else if (add_char(str, arg))
-			return (free(arg->value), -1);
-		j++;
-	}
-	arg->len = ft_strlen(arg->value);
-	return (arg->i);
-}
-
-int	we_have_heredoc(t_expand *arg, char *str)
-{
-	arg->value = ft_strdup("");
-	while (str[arg->i] == '<' || str[arg->i] == ' ')
-		add_char(str, arg);
-	while (str[arg->i] && (!ft_isspace(str[arg->i]) || !(str[arg->i] == '<') || !(str[arg->i] == '>')))
-		add_char(str, arg);
-	return (arg->i);
-}
-
-static int	no_expanding(t_shell *mini, char *str, t_expand *arg, int i)
-{
-	arg->sgl = 0;
-	arg->dbl = 0;
-	arg->i = i;
-	arg->start = i;
-	what_quote(str, arg);
-	while (str[arg->i])
-	{
-		if (((str[arg->i] == ' ' || str[arg->i] == '\t') || (str[arg->i] == '\'' || str[arg->i] == '"')) || arg->dbl || arg->sgl)
-			break ;
-		if (str[arg->i] == '$' && (str[arg->i + 1] == '"' || str[arg->i + 1] == '\'') && !arg->dbl && !arg->sgl)
-			arg->i++;
-		else if (!arg->sgl && !arg->dbl && (str[arg->i] == '\'' || str[arg->i] == '"'))
-		{
-			arg->value = ft_strjoin_char(arg->value, str[arg->i]);
-			what_quote(str, arg);
-		}
-		else if ((arg->sgl && str[arg->i] == '\'') || (arg->dbl && str[arg->i] == '"'))
-		{
-			arg->value = ft_strjoin_char(arg->value, str[arg->i]);
-			what_quote(str, arg);
-		}
-		if ((((arg->dbl && !arg->sgl) || (!arg->dbl && !arg->sgl)) && str[arg->i] == '$' && ((str[arg->i + 1] && ft_isalnum(str[arg->i + 1])) || str[arg->i + 1] == '?')))
-			i = we_have_dollar(mini, arg, str);
-		else if (add_char(str, arg))
-			return (free(arg->value), -1);
-		if (str[arg->i] == '<' && str[arg->i + 1] == '<')
-		{
-			arg->i = we_have_heredoc(arg, str);
-			if (str[arg->i] == '\0')
-				break ;
-		}
-	}
-	arg->len = ft_strlen(arg->value);
-	return (arg->i);
-}
-
-int	arg_no_quotes(t_shell *mini, t_cmd *cmd, t_expand *arg, int i)
-{
-	arg->sgl = 0;
-	arg->dbl = 0;
-	arg->i = i;
-	arg->start = i;
-	(void)mini;
+	the_arg(arg, i);
 	what_quote(cmd->segment, arg);
 	while (cmd->segment[arg->i])
 	{
@@ -146,37 +39,30 @@ int	arg_no_quotes(t_shell *mini, t_cmd *cmd, t_expand *arg, int i)
 	return (arg->i);
 }
 
-int	segment_no_quotes(t_shell *mini, t_cmd *cmd, int i, t_expand *arg)
+int	arg_in_quotes(char *str, int i, t_expand *arg)
 {
-	arg->value = ft_strdup("");
-	if (!arg->value)
+	if (the_arg(arg, i))
 		return (-1);
-	while (cmd->segment[i])
+	what_quote(str, arg);
+	while (str[arg->i])
 	{
-		if ((cmd->segment[i] == '$' && (cmd->segment[i + 1] == '"' || cmd->segment[i + 1] == '\'')) && !arg->dbl && !arg->sgl)
-			i++;
-		if (ft_isspace(cmd->segment[i]) || (cmd->segment[i] == '\'' || cmd->segment[i] == '"'))
+		if (str[arg->i] == '$' && str[arg->i + 1] && (str[arg->i + 1] == '\'' || str[arg->i + 1] == '"') && (!arg->sgl && !arg->dbl))
+			arg->i++;
+		else if (!arg->sgl && !arg->dbl && (str[arg->i] == '\'' || str[arg->i] == '"'))
+			what_quote(str, arg);
+		else if ((arg->sgl && str[arg->i] == '\'') || (arg->dbl && str[arg->i] == '"'))
+			what_quote(str, arg);
+		else if (add_char(str, arg))
+			return (free(arg->value), -1);
+		if (str[arg->i] == ' ' && !arg->sgl && !arg->dbl)
 			break ;
-		if ((cmd->segment[i] == '~' && (ft_isspace(cmd->segment[i + 1]) || cmd->segment[i + 1] == '\0')) || (cmd->segment[i] == '$'
-			&& (cmd->segment[i + 1] && (ft_isalnum(cmd->segment[i + 1])
-			|| cmd->segment[i + 1] == '_' || cmd->segment[i + 1] == '?'))))
-		{
-			arg->i = i;
-			i = expand_variable(mini, cmd->segment, &arg->value, arg);
-		}
-		else
-		{
-			arg->i = i;
-			i = no_expanding(mini, cmd->segment, arg, i);
-		}
 	}
 	arg->len = ft_strlen(arg->value);
-	return (i);
+	return (arg->i);
 }
 
-int	append_to_array(t_cmd *cmd, char *arg, int len, int *index)
+int	append_to_array(t_cmd *cmd, char *arg, int *index)
 {
-	(void)len;
 	cmd->args[*index] = ft_strdup(arg);
 	if (!cmd->args[*index])
 	{
@@ -186,4 +72,3 @@ int	append_to_array(t_cmd *cmd, char *arg, int len, int *index)
 	(*index)++;
 	return (0);
 }
-
