+++ sources/parser/handle_cmd_args_utils2.c
@@ -6,128 +6,74 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/02 09:18:09 by nzharkev          #+#    #+#             */
-/*   Updated: 2024/12/20 14:01:46 by henbuska         ###   ########.fr       */
+/*   Updated: 2024/12/21 16:35:46 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int	handle_question(t_shell *mini, char *str, char **expanded, t_expand *arg)
+int only_redirect(char *str, int i)
 {
-	int	cont;
-	arg->start = arg->i;
-	cont = arg->start;
-	if (str[arg->i] == '$')
-		arg->i = oh_its_a_dollar(mini, str, expanded, arg);
-	arg->i = cont + 2;
-	return (arg->i);
-}
-
-int	new_result(t_expand *arg, char *temp)
-{
-	char	*new_res;
-
-	new_res = ft_strjoin(arg->value, temp);
-	if (!new_res)
-	{
-		free(temp);
-		return (-1);
-	}
-	free(arg->value);
-	free(temp);
-	arg->value = new_res;
-	return (0);
-}
-
-int	we_have_dollar(t_shell *mini, t_expand *arg, char *str)
-{
-	int		s_exp;
-	char	*temp;
-	char	*new_res;
-
-	//printf("str: {%s}\n", str);
-	s_exp = arg->i;
-	temp = ft_strdup("");
-	if (!temp)
-		return (-1);
-	arg->start = arg->i;
-	arg->i = oh_its_a_dollar(mini, str, &temp, arg);
-	if (arg->i == -1)
-	{
-		free(temp);
-		return (-1);
-	}
-	new_res = ft_strjoin(arg->value, temp);
-	if (!new_res)
-	{
-		free(temp);
-		return (-1);
-	}
-	free(arg->value);
-	free(temp);
-	arg->value = new_res;
-	return (0);
-}
-
-void	what_quote(char *str, t_expand *arg)
-{
-	// printf("str[%d]: %c\n", arg->i, str[arg->i]);
-	if ((arg->sgl == 1 && str[arg->i] == '\'') || (arg->dbl == 1
-		&& str[arg->i] == '"'))
-	{
-		if (str[arg->i] == '\'')
-			arg->sgl = !arg->sgl;
-		else if (str[arg->i] == '\"')
-			arg->dbl = !arg->dbl;
-		arg->i++;
-		return ;
-	}
-	if (str[arg->i] == '\'' && arg->sgl == 0)
-	{
-		arg->sgl = 1;
-		arg->i++;
-		// printf("arg->dbl: %d\narg->sgl: %d\n", arg->dbl, arg->sgl);
-		return ;
-	}
-	if (str[arg->i] == '"' && arg->dbl == 0)
-	{
-		arg->dbl = 1;
-		arg->i++;
-		// printf("arg->dbl: %d\narg->sgl: %d\n", arg->dbl, arg->sgl);
-		return ;
-	}
-}
-
-int	the_arg(t_expand *arg, int i)
-{
-	arg->sgl = 0;
-	arg->dbl = 0;
-	arg->i = i;
-	arg->start = i;
-	arg->name = NULL;
-	arg->value = ft_strdup("");
-	if (!arg->value)
-	{
-		ft_putendl_fd("Malloc fail", 2);
+	int n;
+	n = i;
+	while (str[n] && str[n] != '=')
+		n++;
+	if (!str[n])
+		return (0);
+	n++;
+	n++;
+	if ((str[n] == '>' || str[n] == '<') || str[n] == '|')
 		return (1);
+	return (0);
+}
+
+int	init_args_array(t_cmd *cmd, int i)
+{
+	cmd->args = ft_calloc(cmd->args_count = count_args(cmd, i) + 2, sizeof(char *));
+	if (!cmd->args)
+		return (-1);
+	cmd->args[0] = ft_strdup(cmd->command);
+	if (!cmd->args[0])
+	{
+		ft_free_array(cmd->args);
+		return (-1);
 	}
 	return (0);
 }
 
-int	add_char(char *str, t_expand *arg)
+int	count_if_redirection(t_cmd *cmd, int i)
 {
-	char	*temp;
-	char	*temp2;
+	while (cmd->segment[i] && ft_isspace(cmd->segment[i]))
+		i++;
+	while (cmd->segment[i] && !ft_isspace(cmd->segment[i]) &&
+		is_redirection(cmd, i))
+		i++;
+	return (i);
+}
 
-	temp2 = ft_strndup(&str[arg->i], 1);
-	if (!temp2)
-		return (1);
-	temp = ft_strjoin(arg->value, temp2);
-	free(temp2);
-	if (!temp)
-		return (1);
-	free(arg->value);
-	arg->value = temp;
-	arg->i++;
-	return (0);
+int	count_args(t_cmd *cmd, int i)
+{
+	int	args_count;
+
+	args_count = 0;
+	while (cmd->segment[i] && ft_isspace(cmd->segment[i]))
+		i++;
+	while (cmd->segment[i])
+	{
+		if (is_redirection(cmd, i))
+		{
+			i++;
+			i = count_if_redirection(cmd, i);
+		}
+		else
+		{
+			args_count++;
+			while (cmd->segment[i] && !ft_isspace(cmd->segment[i]) &&
+			!is_redirection(cmd, i))
+				i++;
+		}
+		while (cmd->segment[i] && ft_isspace(cmd->segment[i]))
+			i++;
+	}
+	return (args_count);
 }
