+++ sources/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 16:26:26 by henbuska          #+#    #+#             */
-/*   Updated: 2024/12/20 15:48:12 by henbuska         ###   ########.fr       */
+/*   Updated: 2024/12/21 20:12:38 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,10 +28,7 @@ int	parse_and_validate_input(char **input, t_shell *mini)
 	if (split_input_by_pipes(*input, mini))
 		return (1);
 	if (parse_input(mini))
-	{
-		//clean_cmds(mini->cmds);
 		return (1);
-	}
 	return (0);
 }
 
@@ -47,17 +44,6 @@ int	parse_input(t_shell *mini)
 			clean_cmds(mini->cmds);
 			return (1);
 		}
-		/*if (mini->cmds[index]->command)
-		{
-			if (is_this_built(mini->cmds[index]->command) != 1)
-			{
-				if (get_cmd_path(mini, mini->cmds[index]) == 1)
-				{
-					clean_cmds(mini->cmds);
-					return (1);
-				}
-			}
-		}*/
 		index++;
 	}
 	return (0);
@@ -70,8 +56,8 @@ int	parse_cmd_string(t_shell *mini, t_cmd *cmd)
 
 	i = 0;
 	cmd_found = 0;
-	//if (cmd->segment[0] == '"' && cmd->segment[1] == '"')
-	//	return (1);
+	if (cmd->segment[0] == '"' && cmd->segment[1] == '"' && !cmd->segment[2])
+		return (1);
 	if (handle_expand(mini, &cmd))
 		return (1);
 	i = cmd_string_while(mini, cmd, i, &cmd_found);
@@ -94,7 +80,7 @@ int	cmd_string_while(t_shell *mini, t_cmd *cmd, int i, int *cmd_found)
 		}
 		else if (*cmd_found == 0)
 		{
-			i = handle_cmd_name(mini, cmd, i);
+			i = handle_cmd_name(cmd, i);
 			if (i == -1)
 				return (-1);
 			*cmd_found = 1;
