+++ sources/parser/parser_utils.c
@@ -6,7 +6,7 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 17:13:39 by henbuska          #+#    #+#             */
-/*   Updated: 2024/12/18 19:27:34 by henbuska         ###   ########.fr       */
+/*   Updated: 2024/12/21 16:44:32 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,4 +43,25 @@ bool	is_empty_command(t_cmd *cmd, int i)
 		}
 	}
 	return (false);
+}
+
+int	add_char(char *str, t_expand *arg)
+{
+	char	*temp;
+	char	*temp2;
+	char	*temp3;
+
+	temp2 = ft_strndup(&str[arg->i], 1);
+	if (!temp2)
+		return (1);
+	temp3 = ft_strdup(arg->value);
+	temp = ft_strjoin(temp3, temp2);
+	free(temp2);
+	free(temp3);
+	if (!temp)
+		return (1);
+	free(arg->value);
+	arg->value = temp;
+	arg->i++;
+	return (0);
 }
